requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
pandas==2.2.3
pymysql==1.1.1
cryptography==44.0.0
python-dateutil==2.9.0.post0
email-validator==2.2.0
bcrypt==4.2.1
gunicorn==23.0.0
